Refactor default router state to off-link route state
This change prepares for a later change which supports the NDP
Route Information option to discover more-specific routes, as
per RFC 4191.

The newly introduced off-link route state will be used to hold
both the state for default routers (which is a default (off-link)
route through the router, and more-specific routes (which are
routes through some router to some destination subnet more specific
than the IPv6 empty subnet).

Updates #6172.

PiperOrigin-RevId: 378938681